Provide an Add or Update Option
This replaces the current '--replace' option with '--action replace'.

The ation is an option which can be: * Replace * Add (Accrete, Update,
Upsert) * New (the default)

Where '--action add' will overwrite any existing target files, but
otherwise leave the target as is. '--action new' expects the target to
not exist before the run and it will create the target. '--action
replace' is similar to '--action new' but it will first remove the
existing target if it exists.